requirements.txt
Roboflow
hydra-core
torch
ultralytics